Rename _base_ensemble.py to base_ensemble.py
